server/requirements.txt
aiosqlite==0.16.1
APScheduler==3.7.0
bcrypt==3.2.0
bidict==0.21.2
bottle==0.12.19
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
eventlet==0.30.2
fastapi==0.63.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SocketIO==5.0.1
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
gkeepapi==0.13.4
gpsoauth==0.4.3
greenlet==1.0.0
h11==0.12.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.1
idna==2.10
iso8601==0.1.14
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.2
opencv-python==4.5.1.48
pandas==1.2.3
passlib==1.7.4
pigpio==1.78
priority==1.3.0
pycparser==2.20
pycryptodomex==3.10.1
pydantic==1.8.1
pypika-tortoise==0.1.0
python-dateutil==2.8.1
python-engineio==4.0.1
python-mpv==0.5.2
python-multipart==0.0.5
python-socketio==5.1.0
pytz==2020.5
requests==2.25.1
six==1.15.0
starlette==0.13.6
toml==0.10.2
tortoise-orm==0.17.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.11
Werkzeug==1.0.1
wsproto==1.0.0
zope.event==4.5.0
zope.interface==5.3.0
